no doubles

+++ qrs.h
@@ -25,7 +25,7 @@ int peakDetermination(int data);
 void arrayInsert(int* array, int size, int value);
 int* findingPeaks(int data);
 int* findingTime(int data);
-double calculateAverage(int* array);
+int calculateAverage(int* array);
 void recalculateThresholds(QRS_params *params);
 int peakDetection(QRS_params *params, int data);
